+++ helpers/patterns.py
@@ -1,0 +1,18 @@
+
+
+STR_PATTERNS = [
+        r'^[A-Z]+$',
+        r'^[A-Z ]+$',
+        r'^[A-Z#]+$',
+        r'^[a-z]+$',
+        r'^[A-Z0-9]+$',
+        r'^[a-z0-9]+$',
+        r'^[a-zA-Z0-9]+$'
+    ]
+
+DATE_FORMATS = [
+        '%d.%m.%Y',
+        '%m.%d.%Y',
+        '%Y.%m.%d',
+        '%Y.%d.%m',
+    ]
